Disable two songs and add a sneak peek...
The song tags for "PLAY" and "SAIKOU" need an update, because they are as of July 31, 2019 part of Monstercat's new compilation album, Monstercat Uncaged Vol. 7!

Also, "Duality" and "Spection", both by Fractal, have been added to the list as a sneak peek. Stay tuned for their release on Visimusic; they will be here very soon!