Add new song!
"Caps On, Hats Off" by Bossfight has been added to the mix. Also, there is a new option, but it can't be changed yet - once all of the 5th version song tags are up, you can CHOOSE to have either version 4 OR 5 song tags! You better believe it, because it's coming real soon to Visimusic!