Add lyrics for "Till It's Over"
This is the fourth song of 29 that now has timed lyrics. 25 to go!

Also, I have added a new feature: Total Play Counts! I still need to work on the back end of this, but once it's fully functional, you'll be able to track how many times you've played each song to the end, on a song-by-song basis!